requirements.txt
flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.3.1
joblib==1.3.2
numpy==2.2.2
pandas==2.2.3
dnspython==2.8.0
ipwhois==1.3.0
python-whois==0.9.4
requests==2.32.3
python-dotenv==1.0.1
python-telegram-bot==21.10
xgboost>=2.0.0
